requirements.txt
Flask>=2.2.5
requests>=2.32.0
python-dotenv>=1.2.1
Flask-Cors>=6.0.2
pillow>=10.0
google-genai>=0.13.2